requirements.txt
pymongo
insightface
numpy
faiss-cpu
opencv-python-headless
Pillow
streamlit